+++ CHANGELOG.md
@@ -5,8 +5,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.6.1] - 2020-07-01
+### Added
 
+- The build task `Generate_Conceptual_Help` can now remove markdown code
+  from the schema MOF parameter descriptions if markdown code is used to
+  improve the Wiki documentation.
+
+## [0.6.1] - 2020-07-01
+  
 ### Fixed
 
 - Update README.md with the correct build task name; 'Publish_GitHub_Wiki_Content'.